+++ seminar_3/src/model/SaleDTO.java
@@ -1,8 +1,9 @@
 package model;
 
+import java.math.BigDecimal;
 import java.time.LocalTime;
 
 import integration.ItemDTO;
 
-
-public record SaleDTO(int id, LocalTime dateTime, ItemDTO[] boughtItems, float totalPrice, float totalVat, float amountPaid, float change) {}
+public record SaleDTO(int id, LocalTime dateTime, ItemDTO[] boughtItems, BigDecimal totalPrice, BigDecimal totalVat,
+        BigDecimal amountPaid, BigDecimal change) {}
