+++ seminar_3/src/model/SaleInfoDTO.java
@@ -1,5 +1,7 @@
 package model;
 
+import java.math.BigDecimal;
+
 import integration.ItemDTO;
 
-public record SaleInfoDTO(ItemDTO currentItem, double totalPrice, double totalVat) {}
+public record SaleInfoDTO(ItemDTO currentItem, BigDecimal totalPrice, BigDecimal totalVat) {}
